requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.2.0
altair==5.0.1
anndata==0.11.3
array_api_compat==1.11.1
astroid==3.3.8
attrs==24.3.0
blinker==1.9.0
Bottleneck==1.4.2
Brotli==1.0.9
build==1.2.2.post1
cached-property==1.5.2
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.3.2
chex==0.1.88
click==8.1.7
contourpy==1.3.1
cycler==0.11.0
Deprecated==1.2.13
dill==0.3.8
docrep==0.3.2
et-xmlfile==1.1.0
etils==1.12.2
exceptiongroup==1.2.0
filelock==3.13.1
flax==0.10.2
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.7
GitPython==3.1.43
gmpy2==2.2.1
grpcio==1.62.2
h5py==3.13.0
humanize==3.10.0
idna==3.7
importlib_metadata==8.5.0
importlib_resources==6.4.0
isort==5.13.2
jax==0.4.31
jaxlib==0.4.31
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.8
legacy-api-wrap==1.4.1
lightning==2.5.0.post0
lightning-utilities==0.11.9
llvmlite==0.44.0
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mccabe==0.7.0
mdurl==0.1.0
ml_collections==1.0.0
ml_dtypes==0.5.0
mpmath==1.3.0
msgpack==1.0.3
mudata==0.3.1
multidict==6.1.0
multipledispatch==0.6.0
natsort==7.1.1
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.0
numexpr==2.10.1
numpy==1.26.4
numpyro==0.18.0
openpyxl==3.1.5
opentelemetry-api==1.31.0
opt-einsum==3.3.0
optax==0.2.4
orbax-checkpoint==0.6.4
packaging==24.2
pandas==2.2.2
patsy==1.0.1
pillow==11.1.0
pip==25.0
platformdirs==3.10.0
plotly==5.24.1
propcache==0.2.0
protobuf==4.25.3
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.15.1
pylint==3.3.3
pynndescent==0.5.10
pyparsing==3.2.0
pyproject_hooks==1.0.0
pyro-api==0.1.2
pyro-ppl==1.9.1+ab0491a
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytz==2024.1
PyYAML==6.0.2
referencing==0.30.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scanpy==1.11.0
scikit-learn==1.5.2
scipy==1.14.1
scvi-tools==1.3.0
seaborn==0.13.2
session-info2==0.1.2
setuptools==75.8.0
six==1.16.0
smmap==4.0.0
sparse==0.15.4
statsmodels==0.14.4
streamlit==1.43.2
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.17.0
tensorboard_data_server==0.7.0
tensorstore==0.1.65
threadpoolctl==3.5.0
toml==0.10.2
tomlkit==0.13.2
toolz==1.0.0
torch==2.5.1
torchmetrics==1.4.0.post0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2023.3
umap-learn==0.5.7
unicodedata2==15.1.0
urllib3==2.3.0
watchdog==4.0.2
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
xarray==2024.11.0
yarl==1.18.0
zipp==3.21.0
